Created the dev branch
The dev branch is meant for code that is singularly <em>not</em> working in the
least, and is subject to heavy change. I pushed the changes to a
different branch mostly just to see if a push would make the branch in
GitHub, but it's probably a good practice to do anyways.