finished shift thing
added list comprehension, which is cooler than the whole for-loop append
business i was doing. also actually made it work. to decode, just input
the shift number as a negative, which should make it reverse it's
changes. next thing to add is insurance against a problem I think
happens if you try to shift by something greater than 26.